DONE Refactor Infra layer; IN PROGRESS Implementing all other entities;
I've decided to change repositories and ORM models, that were using Domain Entities, to actually use proper Infra models. The reason behind this is basically the hard time to configure EF Core to work with abstract classes. Now I'm using anemic models as a workaround to make my implementations of SOLID work properly; It could be a bad practice in large projects, and the only reason I'm doing this is due the deadline and the low complexity involved.